+++ library/src/main/java/com/pengrad/telegrambot/model/PreparedInlineMessage.kt
@@ -2,5 +2,5 @@ package com.pengrad.telegrambot.model
 
 data class PreparedInlineMessage(
     @get:JvmName("id") val id: String,
-    @get:JvmName("expirationDate") val expirationDate: Long
+    @get:JvmName("expirationDate") val expirationDate: Int
 )
