+++ library/src/main/java/com/pengrad/telegrambot/model/SuccessfulPayment.kt
@@ -4,7 +4,7 @@ data class SuccessfulPayment(
     @get:JvmName("currency") val currency: String,
     @get:JvmName("totalAmount") val totalAmount: Int,
     @get:JvmName("invoicePayload") val invoicePayload: String,
-    @get:JvmName("subscriptionExpirationDate") val subscriptionExpirationDate: Long? = null,
+    @get:JvmName("subscriptionExpirationDate") val subscriptionExpirationDate: Int? = null,
     @get:JvmName("isRecurring") val isRecurring: Boolean? = null,
     @get:JvmName("isFirstRecurring") val isFirstRecurring: Boolean? = null,
     @get:JvmName("shippingOptionId") val shippingOptionId: String? = null,
